new obstacles and slight zombie fix
Added a barrel (axe block): if player presses e/q close to it will 'roll' opposite to the player until it hits something
Added a lever: if player presses e/q next to it will turn on or off certain elements (ie can flip certain blocks to be on or off)
Started a Fall pit(void): if player falls in pit they will 'fall'. check fall function for future improvements needed
made zombie travel faster and more sensitve to player so it looks a bit better but not perfect yet.
Started a spike ball / blade Trap: if player is directly in the path of trap vertically or horizontally the trap will go to hit player, still needs fixes
Enemy sprite flicker: enemies now flash white if they are hit but not dead. best way to see it currently is to use the R projectile as they die in one hit from sword.
Changes are all on demo at Barrel_Lever_Demo scene